Pfuf

+++ lib/GUI_lib/include/tusb_data.hpp
@@ -10,7 +10,7 @@ namespace HID
 
     struct MOUSE_t
     {
-        int8_t mousePos[2] = {};
+        int16_t mousePos[2] = {};
         uint8_t mouseKeys[5] = {};
         uint64_t processTime;
         bool moved;
